main-application/back-end/community-backend-app/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.191.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.65.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
packaging==26.0
proto-plus==1.27.1
protobuf==5.29.6
psycopg2==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
